Moved packet to separate module whilst preserving import paths
